backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
psycopg[binary]==3.1.18
alembic==1.13.1
pydantic[email]==2.7.0
python-jose==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.9
slowapi==0.1.8
geojson==3.1.0
haversine==2.8.0
structlog==24.1.0
httpx==0.27.0
python-dotenv==1.0.1
SQLAlchemy-Utils==0.41.1

# Dev
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
factory-boy==3.3.0
Faker==24.9.0
